moved code that belongs to the binder to a different file
